chore: Extract dev-server status data model
