+++ src/Uno.UI.RemoteControl/RemoteControlClient.cs
@@ -21,6 +21,7 @@ using Uno.UI.RemoteControl.HotReload;
 using Uno.UI.RemoteControl.HotReload.Messages;
 using Uno.UI.RemoteControl.Messages;
 using Windows.Networking.Sockets;
+using static Uno.UI.RemoteControl.RemoteControlStatus;
 
 namespace Uno.UI.RemoteControl;
