+++ src/Uno.UI.RemoteControl/RemoteControlStatus.cs
@@ -1,0 +1,146 @@
+using System;
+using System.Collections.Immutable;
+using System.Linq;
+using System.Text;
+
+namespace Uno.UI.RemoteControl;
+
+internal record RemoteControlStatus(
+	RemoteControlStatus.ConnectionState State,
+	bool? IsVersionValid,
+	(RemoteControlStatus.KeepAliveState State, long RoundTrip) KeepAlive,
+	ImmutableHashSet<RemoteControlStatus.MissingProcessor> MissingRequiredProcessors,
+	(long Count, ImmutableHashSet<Type> Types) InvalidFrames)
+{
+	public (Classification kind, string message) GetSummary()
+	{
+		var (kind, message) = State switch
+		{
+			ConnectionState.Idle => (Classification.Info, "Initializing..."),
+			ConnectionState.NoServer => (Classification.Error, "No server configured, cannot initialize connection."),
+			ConnectionState.Connecting => (Classification.Info, "Connecting to dev-server."),
+			ConnectionState.ConnectionTimeout => (Classification.Error, "Failed to connect to dev-server (timeout)."),
+			ConnectionState.ConnectionFailed => (Classification.Error, "Failed to connect to dev-server (error)."),
+			ConnectionState.Reconnecting => (Classification.Info, "Connection to dev-server has been lost, reconnecting."),
+			ConnectionState.Disconnected => (Classification.Error, "Connection to dev-server has been lost, will retry later."),
+
+			ConnectionState.Connected when IsVersionValid is false => (Classification.Warning, "Connected to dev-server, but version mis-match with client."),
+			ConnectionState.Connected when InvalidFrames.Count is not 0 => (Classification.Warning, $"Connected to dev-server, but received {InvalidFrames.Count} invalid frames from the server."),
+			ConnectionState.Connected when MissingRequiredProcessors is { IsEmpty: false } => (Classification.Warning, "Connected to dev-server, but some required processors are missing on server."),
+			ConnectionState.Connected when KeepAlive.State is KeepAliveState.Late => (Classification.Info, "Connected to dev-server, but keep-alive messages are taking longer than expected."),
+			ConnectionState.Connected when KeepAlive.State is KeepAliveState.Lost => (Classification.Warning, "Connected to dev-server, but last keep-alive messages have been lost."),
+			ConnectionState.Connected when KeepAlive.State is KeepAliveState.Aborted => (Classification.Warning, "Connected to dev-server, but keep-alive has been aborted."),
+			ConnectionState.Connected => (Classification.Ok, "Connected to dev-server."),
+
+			_ => (Classification.Warning, State.ToString()),
+		};
+
+		if (KeepAlive.RoundTrip >= 0)
+		{
+			message += $" (ping {KeepAlive.RoundTrip}ms)";
+		}
+
+		return (kind, message);
+	}
+
+	internal string GetDescription()
+	{
+		var details = new StringBuilder(GetSummary().message);
+
+		if (MissingRequiredProcessors is { Count: > 0 } missing)
+		{
+			details.AppendLine();
+			details.AppendLine();
+			details.AppendLine("Some processor(s) requested by the client are missing on the server:");
+
+			foreach (var m in missing)
+			{
+				details.AppendLine($"- {m.TypeFullName} v{m.Version}: {m.Details}");
+				if (m.Error is not null)
+				{
+					details.AppendLine($"  {m.Error}");
+				}
+			}
+		}
+
+		if (InvalidFrames.Types is { Count: > 0 } invalidFrameTypes)
+		{
+			details.AppendLine();
+			details.AppendLine();
+			details.AppendLine($"Received {InvalidFrames.Count} invalid frames from the server. Failing frame types ({invalidFrameTypes.Count}):");
+
+			foreach (var type in invalidFrameTypes)
+			{
+				details.AppendLine($"- {type.FullName}");
+			}
+		}
+
+		return details.ToString();
+	}
+
+	internal record struct MissingProcessor(string TypeFullName, string Version, string Details, string? Error = null);
+
+	internal enum KeepAliveState
+	{
+		Idle,
+		Ok, // Got ping/pong in expected delays
+		Late, // Sent ping without response within delay
+		Lost, // Got an invalid pong response
+		Aborted // KeepAlive was aborted
+	}
+
+	internal enum ConnectionState
+	{
+		/// <summary>
+		/// Client as not been started yet
+		/// </summary>
+		Idle,
+
+		/// <summary>
+		/// No server information to connect to.
+		/// </summary>
+		NoServer,
+
+		/// <summary>
+		/// Attempting to connect to the server.
+		/// </summary>
+		Connecting,
+
+		/// <summary>
+		/// Reach timeout while trying to connect to the server.
+		/// Connection HAS NOT been established.
+		/// </summary>
+		ConnectionTimeout,
+
+		/// <summary>
+		/// Connection to the server failed.
+		/// Connection HAS NOT been established.
+		/// </summary>
+		ConnectionFailed,
+
+		/// <summary>
+		/// Connection to the server has been established.
+		/// </summary>
+		Connected,
+
+		/// <summary>
+		/// Reconnecting to the server.
+		/// Connection has been established once but lost since then, reconnecting to the SAME server.
+		/// </summary>
+		Reconnecting,
+
+		/// <summary>
+		/// Disconnected from the server.
+		/// Connection has been established once but lost since then and cannot be restored for now but will be retried later.
+		/// </summary>
+		Disconnected
+	}
+
+	internal enum Classification
+	{
+		Ok,
+		Info,
+		Warning,
+		Error
+	}
+}
